cleanup data models in prep to refactor
